+++ SourceCode/Data/Theme.cs
@@ -1,0 +1,8 @@
+﻿namespace ModulesRegistry.Data;
+
+public enum Theme
+{
+    Undefined = 0,
+    Europe = 1,
+    US = 2
+}